some commentary
