+++ pyspecProcScripts/__init__.py
@@ -16,7 +16,6 @@ from .DCCT_func import DCCT
 from .integral_w_error import integral_w_errors, active_propagation
 from .apod_matched_filter import apod_matched_filter
 from .lorentzian_to_gaussian import L2G
-from .calc_var import masked_mean_multi, masked_var_multi
 __all__ = ['calc_baseline',
         'center_echo',
         'correl_align',
@@ -44,7 +43,5 @@ __all__ = ['calc_baseline',
         'zeroth_order_ph',
         'apod_matched_filter',
         'L2G',
-        'masked_mean_multi',
-        'masked_var_multi'
         ]
